Changes to Card Class: Added a toString method. Added a method to compare the magicHand and what the user inputs.
Changes to CardTrick Class: 
Added a random number generator for both the value and suit of card. 
Added user input statements for card value and suit. 
Added a for loop to compare user input to the randomly generate numbers.
Added a print method to let the user know if they won or lost.